requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
Django==2.2.7
django-filter==2.2.0
django-redis-cache==2.1.0
django-rest-framework==0.1.0
django-rest-registration==0.5.2
djangorestframework==3.9.4
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
Pillow==6.2.1
psycopg2==2.8.4
pycrypto==2.6.1
python-redis-lock==3.3.1
pytz==2019.3
redis==3.3.11
redis-structures==0.1.7
six==1.13.0
sqlparse==0.3.0
uWSGI==2.0.18
vine==1.3.0
vital-tools==0.1.13
zipp==0.6.0
